LR-546 code changes against review points

+++ lern-data-products/src/main/scala/org/sunbird/core/util/DataSecurityUtil.scala
@@ -48,23 +48,18 @@ object DataSecurityUtil {
     }
   }
 
-  def getSecuredExhaustFile(jobId: String, orgId: String, channel: String, csvFile: String, encryptionKey: String, storageConfig: StorageConfig, request: JobRequest) (implicit conf: Configuration, fc: FrameworkContext): Unit = {
-    val level = getSecurityLevel(jobId, orgId)
+  def getSecuredExhaustFile(level: String, orgId: String, channel: String, csvFile: String, encryptionKey: String, storageConfig: StorageConfig, request: JobRequest) (implicit conf: Configuration, fc: FrameworkContext): Unit = {
     level match {
       case "L1" =>
-        csvFile
+        zipAndEncrypt("", storageConfig, request, csvFile, level)
       case "L2" =>
-        zipAndEncrypt(csvFile, storageConfig, request)
+        zipAndEncrypt(csvFile, storageConfig, request,"", level)
       case "L3" =>
-        //val downloadPath = Constants.TEMP_DIR + orgId
         val downloadPath = Constants.TEMP_DIR + orgId
         val publicPemFile = httpUtil.downloadFile(encryptionKey, downloadPath)
         encryptionFile(publicPemFile, csvFile)
       case "L4" =>
         val exhaustEncryptionKey = getExhaustEncryptionKey(orgId, channel)
-        //        val exhaustEncryptionKey = "https://sunbirddevbbpublic.blob.core.windows.net/sunbird-content-dev/organisation/0137774123743232000/public.pem"
-        // Download the exhaustEncryptionKey
-        //val downloadPath = Constants.TEMP_DIR + orgId
         val downloadPath = Constants.TEMP_DIR + orgId
         val publicPemFile = httpUtil.downloadFile(exhaustEncryptionKey, downloadPath)
         encryptionFile(publicPemFile, csvFile)
@@ -75,31 +70,33 @@ object DataSecurityUtil {
   }
 
   def getExhaustEncryptionKey(orgId: String, channel: String): String = {
-    val requestBody = Map("request" -> (if(!orgId.isEmpty) Map("organisationId" -> orgId) else Map("channel" -> channel, "isTenant" -> true)))
+      val responseBody = getOrgDetails(orgId, channel)
+      val keys = responseBody.getOrElse("result", Map[String, AnyRef]()).asInstanceOf[Map[String, AnyRef]]
+        .getOrElse("response", Map[String, AnyRef]()).asInstanceOf[Map[String, AnyRef]]
+        .getOrElse("content", List[Map[String, AnyRef]]()).asInstanceOf[List[Map[String, AnyRef]]].head
+        .getOrElse("keys", Map[String, AnyRef]()).asInstanceOf[Map[String, AnyRef]]
+      val exhaustEncryptionKey = keys.getOrElse("exhaustEncryptionKey", List()).asInstanceOf[List[String]]
+      if (exhaustEncryptionKey.nonEmpty) exhaustEncryptionKey.head else ""
+  }
+
+  def getOrgDetails(orgId: String, channel: String): Map[String, AnyRef] = {
+    val requestBody = Map("request" -> (if(!orgId.isEmpty) Map("id" -> orgId) else Map("channel" -> channel, "isTenant" -> true)))
     val request = JSONUtils.serialize(requestBody)
     val headers: Map[String, String] = Map("Content-Type" -> "application/json")
     val httpUtil = new HttpUtil
     val httpResponse = httpUtil.post(Constants.ORG_SEARCH_URL, request, headers)
+    var responseBody = Map[String, AnyRef]().empty
     if (httpResponse.status == 200) {
-      JobLogger.log(s"getOrgDetail for org=$orgId, response body=${httpResponse.body}", None, INFO)(new String())
-      val responseBody = JSONUtils.deserialize[Map[String, AnyRef]](httpResponse.body)
-      val keys = responseBody.getOrElse("result", Map[String, AnyRef]()).asInstanceOf[Map[String, AnyRef]]
-        .getOrElse("response", Map[String, AnyRef]()).asInstanceOf[Map[String, AnyRef]]
-        .getOrElse("keys", Map[String, AnyRef]()).asInstanceOf[Map[String, AnyRef]]
-      val exhaustEncryptionKey = keys.getOrElse("exhaustEncryptionKey", List()).asInstanceOf[List[String]]
-      if (exhaustEncryptionKey.nonEmpty) exhaustEncryptionKey.head else ""
-    } else
-    ""
+      JobLogger.log(s"getOrgDetail for org=$orgId and channel=$channel, response body=${httpResponse.body}", None, INFO)(new String())
+      responseBody = JSONUtils.deserialize[Map[String, AnyRef]](httpResponse.body)
+    }
+    responseBody
   }
 
   @throws(classOf[Exception])
-  private def zipAndEncrypt(url: String, storageConfig: StorageConfig, request: JobRequest)(implicit conf: Configuration, fc: FrameworkContext): String = {
+  private def zipAndEncrypt(url: String, storageConfig: StorageConfig, request: JobRequest, filename: String, level: String)(implicit conf: Configuration, fc: FrameworkContext): Unit = {
 
-    val path = Paths.get(url);
     val storageService = fc.getStorageService(storageConfig.store, storageConfig.accountKey.getOrElse(""), storageConfig.secretKey.getOrElse(""));
-    val tempDir = AppConf.getConfig("spark_output_temp_dir") + request.request_id + "/"
-    val localPath = tempDir + path.getFileName;
-    fc.getHadoopFileUtil().delete(conf, tempDir);
     val filePrefix = storageConfig.store.toLowerCase() match {
       // $COVERAGE-OFF$ Disabling scoverage
       case "s3" =>
@@ -112,37 +109,61 @@ object DataSecurityUtil {
       case _ =>
         storageConfig.fileName
     }
-    val objKey = url.replace(filePrefix, "");
-    if (storageConfig.store.equals("local")) {
-      fc.getHadoopFileUtil().copy(filePrefix, localPath, conf)
+    var objKey = ""
+    var localPath = ""
+    var tempDir = ""
+    var resultFile = ""
+    if(level == "L2") {
+      tempDir = AppConf.getConfig("spark_output_temp_dir") + request.request_id + "/"
+      val path = Paths.get(url);
+      objKey = url.replace(filePrefix, "");
+      localPath = tempDir + path.getFileName;
+      fc.getHadoopFileUtil().delete(conf, tempDir);
+      if (storageConfig.store.equals("local")) {
+        fc.getHadoopFileUtil().copy(filePrefix, localPath, conf)
+      }
+      // $COVERAGE-OFF$ Disabling scoverage
+      else {
+        storageService.download(storageConfig.container, objKey, tempDir, Some(false));
+      }
+    } else {
+      //filePath = "declared_user_detail/"
+      localPath = filename;
+      objKey = localPath.replace(filePrefix, "");
+
     }
-    // $COVERAGE-OFF$ Disabling scoverage
-    else {
-      storageService.download(storageConfig.container, objKey, tempDir, Some(false));
-    }
+
     // $COVERAGE-ON$
     val zipPath = localPath.replace("csv", "zip")
     val zipObjectKey = objKey.replace("csv", "zip")
-    val zipLocalObjKey = url.replace("csv", "zip")
-
-    request.encryption_key.map(key => {
-      val zipParameters = new ZipParameters();
-      zipParameters.setEncryptFiles(true);
-      zipParameters.setEncryptionMethod(EncryptionMethod.ZIP_STANDARD); // AES encryption is not supported by default with various OS.
-      val zipFile = new ZipFile(zipPath, key.toCharArray());
-      zipFile.addFile(localPath, zipParameters)
-    }).getOrElse({
+    if (level == "L2") {
+      val zipLocalObjKey = url.replace("csv", "zip")
+      request.encryption_key.map(key => {
+        val zipParameters = new ZipParameters();
+        zipParameters.setEncryptFiles(true);
+        zipParameters.setEncryptionMethod(EncryptionMethod.ZIP_STANDARD); // AES encryption is not supported by default with various OS.
+        val zipFile = new ZipFile(zipPath, key.toCharArray());
+        zipFile.addFile(localPath, zipParameters)
+      }).getOrElse({
+        new ZipFile(zipPath).addFile(new File(localPath));
+      })
+      resultFile = if (storageConfig.store.equals("local")) {
+        fc.getHadoopFileUtil().copy(zipPath, zipLocalObjKey, conf)
+      }
+      // $COVERAGE-OFF$ Disabling scoverage
+      else {
+        storageService.upload(storageConfig.container, zipPath, zipObjectKey, Some(false), Some(0), Some(3), None);
+      }
+      // $COVERAGE-ON$
+      fc.getHadoopFileUtil().delete(conf, tempDir);
+      resultFile;
+    } else {
       new ZipFile(zipPath).addFile(new File(localPath));
-    })
-    val resultFile = if (storageConfig.store.equals("local")) {
-      fc.getHadoopFileUtil().copy(zipPath, zipLocalObjKey, conf)
+       if (!storageConfig.store.equals("local")) {
+         resultFile = storageService.upload(storageConfig.container, zipPath, zipObjectKey, Some(false), Some(0), Some(3), None)
+      }
+      fc.getHadoopFileUtil().delete(conf, localPath);
+      resultFile
     }
-    // $COVERAGE-OFF$ Disabling scoverage
-    else {
-      storageService.upload(storageConfig.container, zipPath, zipObjectKey, Some(false), Some(0), Some(3), None);
-    }
-    // $COVERAGE-ON$
-    fc.getHadoopFileUtil().delete(conf, tempDir);
-    resultFile;
   }
 }
